requirements.txt
torchvision==0.11.2
pillow
numpy
cv2